requirements.txt
boto3==1.34
PyYAML==6.0.1
requests==2.32.3
SQLAlchemy==2.0.31
PyMySQL==1.1.1